v1.3 - close #26 #31
