+++ ParamList.go
@@ -10,11 +10,26 @@ import (
 // ParamList ..
 type ParamList []*Param
 
-// StructParams - load params from given any struct
+// Get method returns the value associated with the given name
+func (plist ParamList) Get(key string) any {
+	for _, p := range plist {
+		if p.Key == key {
+			return p.Value
+		}
+	}
+	return nil
+}
+
+// Len ..
+func (p ParamList) Len() int {
+	return len(p)
+}
+
+// AnyToParams - load params from given any struct
 // 1) Convert struct to JSON
-// 2) Now convert JSON to map[string]interface{}
+// 2) Now convert JSON to map[string]any
 // 3) Clear params from nil
-func StructParams(v interface{}) ParamList {
+func AnyToParams(v any) ParamList {
 	// to JSON output
 	buf, _ := json.MarshalIndent(v, "", "\t")
 	return JSONToParams(buf)
@@ -23,7 +38,7 @@ func StructParams(v interface{}) ParamList {
 // JSONToParams - load params from JSON
 func JSONToParams(buf []byte) ParamList {
 	// to map
-	m := map[string]interface{}{}
+	m := map[string]any{}
 	if err := json.Unmarshal(buf, &m); err != nil {
 		log.Printf("JSONToParams: %s", err)
 		return nil
@@ -37,17 +52,17 @@ func JSONToParams(buf []byte) ParamList {
 	return params
 }
 
-// walk map[string]interface{} and collect valid params
-func mapToParams(m map[string]interface{}) ParamList {
+// walk map[string]any and collect valid params
+func mapToParams(m map[string]any) ParamList {
 	var params ParamList
 	for mKey, mVal := range m {
 		p := NewParam(mKey)
 
 		switch v := mVal.(type) {
-		case map[string]interface{}:
+		case map[string]any:
 			p.Type = StructParam
 			p.Params = mapToParams(v)
-		case []interface{}:
+		case []any:
 			p.Type = SliceParam
 			p.Params = sliceToParams(v)
 		default:
@@ -66,7 +81,7 @@ func mapToParams(m map[string]interface{}) ParamList {
 }
 
 // sliceToParams - slice of unknown - simple slice or complex
-func sliceToParams(arr []interface{}) ParamList {
+func sliceToParams(arr []any) ParamList {
 	var params ParamList
 
 	for i, val := range arr {
@@ -75,7 +90,7 @@ func sliceToParams(arr []interface{}) ParamList {
 		p := NewParam(i + 1)
 
 		switch v := val.(type) {
-		case map[string]interface{}:
+		case map[string]any:
 			p.Type = StructParam
 			p.Params = mapToParams(v)
 		default:
@@ -93,7 +108,7 @@ func sliceToParams(arr []interface{}) ParamList {
 }
 
 // StructToParams - walk struct and collect valid params
-func StructToParams(paramStruct interface{}) ParamList {
+func StructToParams(paramStruct any) ParamList {
 	var params ParamList
 	var keys reflect.Type
 	var vals reflect.Value
@@ -158,12 +173,12 @@ func reflectStructToParams(p *Param, val reflect.Value) {
 		}
 		p.Type = ImageParam
 		p.SetValue(imgVal)
-	} else {
-		p.Type = StructParam
-		p.Params = StructToParams(val)
+		return
 	}
 
-	return
+	p.Type = StructParam
+	p.Params = StructToParams(val)
+
 }
 
 // reflectSliceToParams - map slice of reflect to params
@@ -198,8 +213,6 @@ func reflectSliceToParams(p *Param, val reflect.Value) {
 
 		p.Params = append(p.Params, itemParam)
 	}
-
-	return
 }
 
 // Parse row content to param list
@@ -207,9 +220,11 @@ func rowParams(row []byte) ParamList {
 	// extract from raw contents
 	re := regexp.MustCompile(ParamPattern)
 	matches := re.FindAllSubmatch(row, -1)
+
 	if matches == nil || matches[0] == nil {
 		return nil
 	}
+
 	var list []*Param
 	for _, match := range matches {
 		p := NewParam(string(match[2]))
