+++ Template.go
@@ -3,7 +3,6 @@ package docxplate
 import (
 	"archive/zip"
 	"bytes"
-	"encoding/xml"
 	"fmt"
 	"io"
 	"log"
@@ -126,51 +125,6 @@ func OpenTemplateWithURL(docurl string) (tpl *Template, err error) {
 	return
 }
 
-// Convert given bytes to struct of xml nodes
-func (t *Template) bytesToXMLStruct(buf []byte) *xmlNode {
-	// Do not strip <w: entiraly, but keep reference as w-t
-	// So any string without w: would stay same, but all w- will be replaced again
-	buf = bytes.ReplaceAll(buf, []byte("<w:"), []byte("<w-"))
-	buf = bytes.ReplaceAll(buf, []byte("</w:"), []byte("</w-"))
-	buf = bytes.ReplaceAll(buf, []byte("<v:"), []byte("<v-"))
-	buf = bytes.ReplaceAll(buf, []byte("</v:"), []byte("</v-"))
-
-	xdocNode := &xmlNode{}
-	if err := xml.Unmarshal(buf, &xdocNode); err != nil {
-		log.Printf("fileToXMLStruct: %v", err)
-	}
-
-	xdocNode.FixNamespaceDuplication()
-	// Assign parent nodes to all nodes
-	xdocNode.Walk(func(xnode *xmlNode) {
-		xnode.FixNamespaceDuplication()
-
-		if xnode.Tag() == "w-body" {
-			xnode.parent = xdocNode
-		}
-
-		for _, n := range xnode.Nodes {
-			n.parent = xnode
-		}
-	})
-
-	// log.Printf("%s", structToXMLBytes(n))
-	return xdocNode
-}
-
-// Convert given file (from template.Files) to struct of xml nodes
-func (t *Template) fileToXMLStruct(fname string) *xmlNode {
-	f, ok := t.files[fname]
-	if !ok {
-		return nil
-	}
-
-	fr, _ := f.Open()
-	buf := readerBytes(fr)
-
-	return t.bytesToXMLStruct(buf)
-}
-
 // Expand some placeholders to enable row replacer replace them
 // Note: Currently only struct type support image replacement
 // Users: []User{ User{Name:AAA}, User{Name:BBB} }
@@ -180,7 +134,7 @@ func (t *Template) fileToXMLStruct(fname string) *xmlNode {
 
 // Params  - replace template placeholders with params
 // "Hello {{ Name }}!"" --> "Hello World!""
-func (t *Template) Params(v interface{}) {
+func (t *Template) Params(v any) {
 	// t.params = collectParams("", v)
 	switch val := v.(type) {
 	case map[string]interface{}:
@@ -193,7 +147,8 @@ func (t *Template) Params(v interface{}) {
 		if reflect.ValueOf(v).Kind() == reflect.Struct {
 			t.params = StructToParams(val)
 		} else {
-			t.params = StructParams(val)
+			// any other type try to convert
+			t.params = AnyToParams(val)
 		}
 	}
 
@@ -205,9 +160,9 @@ func (t *Template) Params(v interface{}) {
 
 			xnode := t.fileToXMLStruct(f.Name)
 
-			// Enchance some markup (removed when building XML in the end)
+			// Enhance some markup (removed when building XML in the end)
 			// so easier to find some element
-			t.enchanceMarkup(xnode)
+			t.enhanceMarkup(xnode)
 
 			// While formating docx sometimes same style node is split to
 			// multiple same style nodes and different content
@@ -226,328 +181,14 @@ func (t *Template) Params(v interface{}) {
 			// otherwise they are left
 			t.triggerMissingParams(xnode)
 
-			// xnode.Walk(func(n *xmlNode) {
-			// 	if is, _ := n.IsListItem(); is {
-			// 		n.Walk(func(wt *xmlNode) {
-			// 			if wt.Tag() == "w-t" {
-			// 				color.Yellow("%s", wt)
-			// 			}
-			// 		})
-			// 	}
-			// })
+			// After all done with placeholders, modify contents
+			// - new lines to docx new lines
+			t.enhanceContent(xnode)
 
 			// Save []bytes
 			t.modified[f.Name] = structToXMLBytes(xnode)
 		}
 	}
-}
-
-// Collect and trigger placeholders with trigger but unset in `t.params`
-// Placeholders with trigger `:empty` must be triggered
-// otherwise they are left
-func (t *Template) triggerMissingParams(xnode *xmlNode) {
-	if t.params == nil {
-		return
-	}
-
-	var triggerParams ParamList
-
-	xnode.Walk(func(n *xmlNode) {
-		if !n.isRowElement() || !n.HaveParams() {
-			return
-		}
-
-		p := NewParamFromRaw(n.AllContents())
-		if p != nil && p.Trigger != nil {
-			triggerParams = append(triggerParams, p)
-		}
-	})
-
-	if triggerParams == nil {
-		return
-	}
-
-	// make sure not to "tint" original t.params
-	_params := t.params
-	t.params = triggerParams
-
-	// do stuff only with filtered params
-	t.replaceSingleParams(xnode, true)
-
-	// back to original
-	t.params = _params
-}
-
-type placeholderType int8
-
-const (
-	singlePlaceholder placeholderType = iota
-	inlinePlaceholder
-	rowPlaceholder
-)
-
-type placeholder struct {
-	Type         placeholderType
-	Placeholders []string
-	Separator    string
-}
-
-// Expand complex placeholders
-func (t *Template) expandPlaceholders(xnode *xmlNode) {
-	t.params.Walk(func(p *Param) {
-		if p.Type != SliceParam {
-			return
-		}
-
-		prefixes := []string{
-			p.PlaceholderPrefix(),
-			p.ToCompact(p.PlaceholderPrefix()),
-		}
-
-		var max int
-		for _, prefix := range prefixes {
-			xnode.Walk(func(nrow *xmlNode) {
-				if nrow.isNew {
-					return
-				}
-				if !nrow.isRowElement() {
-					return
-				}
-				if !nrow.AnyChildContains([]byte(prefix)) {
-					return
-				}
-
-				contents := nrow.AllContents()
-				rowParams := rowParams(contents)
-				rowPlaceholders := make(map[string]*placeholder)
-				// Collect placeholder that for expansion
-				for _, rowParam := range rowParams {
-					var placeholderType placeholderType
-					if len(rowParam.Separator) > 0 {
-						placeholderType = inlinePlaceholder
-					} else {
-						placeholderType = rowPlaceholder
-					}
-
-					var trigger string
-					if rowParam.Trigger != nil {
-						trigger = " " + rowParam.Trigger.String()
-					}
-
-					var isMatch bool
-					var index = -1
-					currentLevel := p.Level
-					placeholders := make([]string, 0, len(p.Params))
-					p.WalkFunc(func(p *Param) {
-						if p.Level == currentLevel+1 {
-							index++
-						}
-						if rowParam.AbsoluteKey == p.CompactKey {
-							isMatch = true
-							placeholders = append(placeholders, "{{"+p.AbsoluteKey+trigger+"}}")
-						}
-					})
-
-					if isMatch {
-						rowPlaceholders[rowParam.RowPlaceholder] = &placeholder{
-							Type:         placeholderType,
-							Placeholders: placeholders,
-							Separator:    strings.TrimLeft(rowParam.Separator, " "),
-						}
-
-						if max < len(placeholders) {
-							max = len(placeholders)
-						}
-					}
-				}
-				// Expand placeholder exactly
-				nnews := make([]*xmlNode, max, max)
-				for oldPlaceholder, newPlaceholder := range rowPlaceholders {
-					switch newPlaceholder.Type {
-					case inlinePlaceholder:
-						nrow.Walk(func(n *xmlNode) {
-							if !inSlice(n.XMLName.Local, []string{"w-t"}) || len(n.Content) == 0 {
-								return
-							}
-							n.Content = bytes.ReplaceAll(n.Content, []byte(oldPlaceholder), []byte(strings.Join(newPlaceholder.Placeholders, newPlaceholder.Separator)))
-						})
-					case rowPlaceholder:
-						defer func() {
-							nrow.delete()
-						}()
-						for i, placeholder := range newPlaceholder.Placeholders {
-							if nnews[i] == nil {
-								nnews[i] = nrow.cloneAndAppend()
-							}
-							nnews[i].Walk(func(n *xmlNode) {
-								if !inSlice(n.XMLName.Local, []string{"w-t"}) || len(n.Content) == 0 {
-									return
-								}
-								n.Content = bytes.ReplaceAll(n.Content, []byte(oldPlaceholder), []byte(placeholder))
-							})
-						}
-					}
-				}
-			})
-		}
-	})
-
-	// Cloned nodes are marked as new by default.
-	// After expanding mark as old so next operations doesn't ignore them
-	xnode.Walk(func(n *xmlNode) {
-		n.isNew = false
-	})
-}
-
-// Replace single params by type
-func (t *Template) replaceSingleParams(xnode *xmlNode, triggerParamOnly bool) {
-	xnode.Walk(func(n *xmlNode) {
-		if n == nil || n.isDeleted {
-			return
-		}
-
-		// node params
-		t.params.Walk(func(p *Param) {
-			for i, attr := range n.Attrs {
-				if strings.Contains(attr.Value, "{{") {
-					n.Attrs[i].Value = string(p.replaceIn([]byte(attr.Value)))
-				}
-			}
-		})
-
-		// node contentt
-		if bytes.Contains(n.Content, []byte("{{")) {
-			// Try to replace on node that contains possible placeholder
-			t.params.Walk(func(p *Param) {
-				// Only string and image param to replace
-				if p.Type != StringParam && p.Type != ImageParam {
-					return
-				}
-				// Prefix check
-				if !n.ContentHasPrefix(p.PlaceholderPrefix()) {
-					return
-				}
-				// Trigger: does placeholder have trigger
-				if p.Trigger = p.extractTriggerFrom(n.Content); p.Trigger != nil {
-					defer func() {
-						p.RunTrigger(n)
-					}()
-				}
-				if triggerParamOnly {
-					return
-				}
-				// Repalce by type
-				switch p.Type {
-				case StringParam:
-					t.replaceTextParam(n, p)
-				case ImageParam:
-					t.replaceImageParams(n, p)
-				}
-			})
-		}
-	})
-}
-
-// wrapper for simple param replace func
-func (t *Template) replaceTextParam(xnode *xmlNode, param *Param) {
-	xnode.Content = param.replaceIn(xnode.Content)
-}
-
-// Image placeholder replace
-func (t *Template) replaceImageParams(xnode *xmlNode, param *Param) {
-	// Sometime the placeholder is in the before or middle of the text, but node is appended in the last.
-	// So, we have to split the text and image into different nodes to achieve cross-display.
-	contentSlice := bytes.Split(xnode.Content, []byte(param.Placeholder()))
-	for i, content := range contentSlice {
-		// text node
-		if len(content) != 0 {
-			contentNode := &xmlNode{
-				XMLName: xml.Name{Space: "", Local: "w-t"},
-				Content: content,
-				parent:  xnode.parent,
-				isNew:   true,
-			}
-			xnode.parent.Nodes = append(xnode.parent.Nodes, contentNode)
-		}
-		// image node
-		if len(contentSlice)-i > 1 {
-			imgNode := t.bytesToXMLStruct([]byte(param.Value))
-			imgNode.parent = xnode.parent
-			xnode.parent.Nodes = append(xnode.parent.Nodes, imgNode)
-		}
-	}
-	// Empty the content before deleting to prevent reprocessing when params walk
-	xnode.Content = []byte("")
-	xnode.delete()
-}
-
-// Enchance some markup (removed when building XML in the end)
-// so easier to find some element
-func (t *Template) enchanceMarkup(xnode *xmlNode) {
-
-	// List items - add list item node `w-p` attributes
-	// so it's recognized as listitem
-	xnode.Walk(func(n *xmlNode) {
-		if n.Tag() != "w-p" {
-			return
-		}
-
-		isListItem, listID := n.IsListItem()
-		if !isListItem {
-			return
-		}
-
-		// n.XMLName.Local = "w-item"
-		n.Attrs = append(n.Attrs, xml.Attr{
-			Name:  xml.Name{Local: "list-id"},
-			Value: listID,
-		})
-
-	})
-}
-
-// This func is fixing broken placeholders by merging "w-t" nodes.
-// "w-p" (Record) can hold multiple "w-r". And "w-r" holts "w-t" node
-// -
-// If these nodes not fixed than params replace can not be done as
-// replacer process nodes one by one
-func (t *Template) fixBrokenPlaceholders(xnode *xmlNode) {
-	xnode.Walk(func(xnode *xmlNode) {
-		if !xnode.isRowElement() {
-			// broken placeholders are in row elements
-			return
-		}
-
-		if !xnode.HaveParams() {
-			// whole text doesn't hold any params
-			return
-		}
-
-		var isBrokenLeftPlaceholder bool
-		var isBrokenRightPlaceholder bool
-		contents := xnode.AllContents()
-		xnode.Walk(func(xnode *xmlNode) {
-			if xnode.Content == nil || len(xnode.Content) == 0 {
-				return
-			}
-			// Match right }} to sub or delete
-			if isBrokenLeftPlaceholder {
-				isBrokenRightPlaceholder = t.matchBrokenRightPlaceholder(string(xnode.Content))
-				if isBrokenRightPlaceholder {
-					xnode.Content = xnode.Content[bytes.Index(xnode.Content, []byte("}}"))+2:]
-				} else {
-					xnode.delete()
-					return
-				}
-			}
-			// Match left {{  to fix broken
-			isBrokenLeftPlaceholder = t.matchBrokenLeftPlaceholder(string(xnode.Content))
-			if isBrokenLeftPlaceholder {
-				xnode.Content = append(xnode.Content, contents[bytes.Index(contents, xnode.Content)+len(xnode.Content):bytes.Index(contents, []byte("}}"))+2]...)
-			}
-			contents = contents[bytes.Index(contents, xnode.Content)+len(xnode.Content):]
-		})
-	})
 }
 
 // Bytes - create docx archive but return only bytes of it
@@ -633,44 +274,6 @@ func (t *Template) Placeholders() []string {
 	arr = re.FindAllString(plaintext, -1)
 
 	return arr
-}
-
-// Match left part placeholder `{{`
-func (t *Template) matchBrokenLeftPlaceholder(content string) bool {
-	stack := make([]string, 0)
-
-	for i, char := range content {
-		if i > 0 {
-			if char == '{' && content[i-1] == '{' {
-				stack = append(stack, "{{")
-			} else if char == '}' && content[i-1] == '}' && len(stack) > 0 {
-				stack = stack[:len(stack)-1]
-			}
-		}
-	}
-
-	return len(stack) > 0
-}
-
-// Match right placeholder part `}}`
-func (t *Template) matchBrokenRightPlaceholder(content string) bool {
-	stack := make([]string, 0)
-
-	for i, char := range content {
-		if i > 0 {
-			if char == '{' && content[i-1] == '{' {
-				stack = append(stack, "{{")
-			} else if char == '}' && content[i-1] == '}' {
-				if len(stack) > 0 {
-					stack = stack[:len(stack)-1]
-				} else {
-					return true
-				}
-			}
-		}
-	}
-
-	return false
 }
 
 // Plaintext - return as plaintext
