+++ go.mod
@@ -2,10 +2,4 @@ module github.com/bobiverse/docxplate
 
 go 1.22.4
 
-require github.com/fatih/color v1.17.0
-
-require (
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	golang.org/x/sys v0.21.0 // indirect
-)
+require github.com/logrusorgru/aurora/v4 v4.0.0
