+++ test-data/user.template-with-images.docx
