+++ version.txt
@@ -1,1 +1,1 @@
-v1.2.4
+v1.3
